chore: update auth page
